pyproject.toml
[project]
name = "search"
version = "0.1.0"
description = "building a search engine over custom data using Vespa"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.4.1",
    "pandas>=2.3.3",
    "pyvespa>=0.62.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "sentence-transformers>=5.1.2",
]
